+++ package.json
@@ -18,6 +18,7 @@
     "@radix-ui/react-separator": "^1.1.7",
     "@radix-ui/react-slot": "^1.2.3",
     "@radix-ui/react-tooltip": "^1.2.7",
+    "@redux-devtools/extension": "^3.3.0",
     "@tailwindcss/postcss": "^4.1.7",
     "axios": "^1.9.0",
     "class-variance-authority": "^0.7.1",
@@ -29,7 +30,8 @@
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "tailwind-merge": "^3.3.1",
-    "tailwindcss": "^4.1.7"
+    "tailwindcss": "^4.1.7",
+    "zustand": "^5.0.5"
   },
   "devDependencies": {
     "@types/node": "^20",
