+++ src/types/auth.ts
@@ -2,3 +2,5 @@ export type AuthResponse = {
   token: string;
   refreshToken: string;
 };
+
+export type User = { name: string };
